Move org, token auth to own files
