+++ app/controllers/ReverseProxy.scala
@@ -1,11 +1,8 @@
 package controllers
 
 import akka.actor.ActorSystem
-import io.flow.common.v0.models.Environment
 import io.flow.token.v0.{Client => TokenClient}
 import io.flow.organization.v0.{Client => OrganizationClient}
-import io.flow.organization.v0.models.{OrganizationAuthorizationForm}
-import io.flow.token.v0.models.{TokenAuthenticationForm, TokenReference}
 import java.util.UUID
 import javax.inject.{Inject, Singleton}
 
@@ -20,24 +17,33 @@ class ReverseProxy @Inject () (
   system: ActorSystem,
   authorizationParser: AuthorizationParser,
   config: Config,
-  flowAuth: FlowAuth,
+  override val flowAuth: FlowAuth,
   proxyConfigFetcher: ProxyConfigFetcher,
   apidocServicesFetcher: ApidocServicesFetcher,
   serverProxyFactory: ServerProxy.Factory,
   ws: play.api.libs.ws.WSClient
-) extends Controller with lib.Errors {
+) extends Controller
+  with lib.Errors
+  with auth.OrganizationAuth
+  with auth.TokenAuth {
 
   val index: Index = proxyConfigFetcher.current()
 
   private[this] implicit val ec = system.dispatchers.lookup("reverse-proxy-context")
 
-  private[this] val organizationClient = {
+  override val organizationClient = {
     val server = mustFindServerByName("organization")
     Logger.info(s"Creating OrganizationClient w/ baseUrl[${server.host}]")
     new OrganizationClient(ws, baseUrl = server.host)
   }
 
-  private[this] val tokenClient = {
+  private[this] val sessionClient = {
+    val server = mustFindServerByName("session")
+    Logger.info(s"Creating SessionClient w/ baseUrl[${server.host}]")
+    new OrganizationClient(ws, baseUrl = server.host) // TODO
+  }
+
+  override val tokenClient = {
     val server = mustFindServerByName("token")
     Logger.info(s"Creating TokenClient w/ baseUrl[${server.host}]")
     new TokenClient(ws, baseUrl = server.host)
@@ -183,7 +189,7 @@ class ReverseProxy @Inject () (
       }
 
       case Some(t) => {
-        resolveOrganizationAuthorization(t, organization).flatMap {
+        resolveOrganization(t, organization).flatMap {
           case None => Future(
             request.response(422, s"Not authorized to access $organization or the organization does not exist")
           )
@@ -242,7 +248,7 @@ class ReverseProxy @Inject () (
       }
     }
   }
-  
+
   /**
     * Resolves the incoming method and path to a specific operation. Also implements
     * overrides from incoming request headers:
@@ -294,7 +300,7 @@ class ReverseProxy @Inject () (
           )
 
           case Some(t) => {
-            resolveOrganizationAuthorization(t, Constants.FlowOrganizationId).map {
+            resolveOrganization(t, Constants.FlowOrganizationId).map {
               case None => {
                 Left(
                   request.response(401, s"Not authorized to access organization[${Constants.FlowOrganizationId}]")
@@ -367,83 +373,6 @@ class ReverseProxy @Inject () (
   private[this] def mustFindServerByName(name: String): Server = {
     findServerByName(name).getOrElse {
       sys.error(s"There is no server named '$name' in the current config: " + index.config.sources.map(_.uri))
-    }
-  }
-
-  /**
-    * Queries token server to check if the specified token is a known
-    * valid token.
-    */
-  private[this] def resolveToken(requestId: String, token: String): Future[Option[TokenReference]] = {
-    tokenClient.tokens.postAuthentications(
-      TokenAuthenticationForm(token = token),
-      requestHeaders = Seq(
-        Constants.Headers.FlowRequestId -> requestId
-      )
-    ).map { tokenReference =>
-      Some(tokenReference)
-
-    }.recover {
-      case io.flow.token.v0.errors.UnitResponse(404) => {
-        None
-      }
-
-      case ex: Throwable => {
-        sys.error(s"Could not communicate with token server at[${tokenClient.baseUrl}]: $ex")
-      }
-    }
-  }
-
-  /**
-    * Queries organization server to authorize this user for this
-    * organization and also pulls the organization's environment.
-    */
-  private[this] def resolveOrganizationAuthorization(
-    token: ResolvedToken,
-    organization: String
-  ): Future[Option[ResolvedToken]] = {
-    val authFuture = (token.environment, token.organizationId) match {
-
-      case (Some(env), Some(orgId)) => {
-        organizationClient.organizationAuthorizations.post(
-          OrganizationAuthorizationForm(
-            organization = organization,
-            environment = Environment(env)
-          ),
-          requestHeaders = flowAuth.headers(token)
-        )
-      }
-
-      case (_, _) => {
-        organizationClient.organizationAuthorizations.getByOrganization(
-          organization = organization,
-          requestHeaders = flowAuth.headers(token)
-        )
-      }
-    }
-
-    authFuture.map { orgAuth =>
-      Some(
-        token.copy(
-          organizationId = Some(organization),
-          environment = Some(orgAuth.environment.toString),
-          role = Some(orgAuth.role.toString)
-        )
-      )
-    }.recover {
-      case io.flow.organization.v0.errors.UnitResponse(401) => {
-        Logger.warn(s"Token[$token] was not authorized for organization[$organization]")
-        None
-      }
-
-      case io.flow.organization.v0.errors.UnitResponse(404) => {
-        Logger.warn(s"Token[$token] organization[$organization] not found")
-        None
-      }
-
-      case ex: Throwable => {
-        sys.error(s"Error communicating with organization server at[${organizationClient.baseUrl}]: ${ex.getMessage}")
-      }
     }
   }
